Distribute make_sanity_checks between cli and api
If all the sanity checks are performed in the context of the cli,
other programs that use the PyScaffold Python API cannot benefit from
it. This issue can be solved by moving some of the checks to another
functions that usually would be part of the public API
(create_project, get_default_opts).