Improve test fixtures
Prior to this commit, the git fixture was not similar enough to the
original function, preventing the logs to be reported and therefore
making it difficult to test logging.

Additionally, this commit updates fixtures to use when possible
`monkeypatch` included in pytest, instead of manually patching.  One of
the advantages is that changes are automatically reversed after the test
function is executed.