Workaround travis error from test-runner using tox
For some reason running `python setup.py test` results in ImportError
on Travis. It seems that the installed pyscaffold is in a older version
(without some submodules).  But the test-runner is supposed to run
inside an isolated env and install the current version of the
package... It does not make any sense...

Let's run tests inside tox (that also use an isolated env).