Conciliate properly changes with Venv (tests)
A subclass of `pytest_virtualv.VirtalEnv` is proposed.
This class updates `pip`, `setuptools` and `certifi` (if necessary), so we
avoid some class of errors.
Moreover the method run is reminiscent from `conftest.py:venv_run` +
some printing for debugging purposes.

The `venv` fixture in `conftest.py` explicitly sets the workspace to be inside
a temporary directory, the current python interpreter (avoid 2.7 if
accessible), and the name `.venv` (avoid Pipenv conflicts).