Introduce global config
As discussed in #236, this commit introduces the ability of storing your
"favourite" configuration for PyScaffold in a config file in the user's
home directory:

- OSX: ~/Library/Preferences/pyscaffold/default.cfg (or something like that)
- Windows: C:\Users\<username>\AppData\Local\pyscaffold\pyscaffold\default.cfg
- Linux/Others: $XDG_CONFIG_HOME/pyscaffold/default.cfg with fallback to ~/.config/pyscaffold/default.cfg

In order to provide these nice and compliant defaults, this commit
introduces a new vendorised package (`appdirs`) inside pyscaffold.contrib.
This is done to avoid adding dependencies to PyScaffold itself.
The `appdirs` module is very small and MIT-licensed, but rewriting its
features would be painful and require lots of efforts, specially for
testing. I assume that this feature is highly valuable for PyScaffold's
users (it gets really tedious after a while to type the same extensions
all the time), and therefore it is worthy.

There is an on going discussion about simplifying PyScaffold to not be
required in build time and therefore eliminating the needs for
vendorising dependencies. With the current written code, that is very
easy to change.

Future works not included in this PR:
- Add a `--config` flag to the CLI to specify passing a different config
  file/ `NO_CONFIG`.

Closes #236.