Clarify documentation on global config
The previous doc used the concept of `preset` to differentiate package
config and PyScaffold's own config. This commit tries to clarify the
difference and remove the confusing term.

Additionally we also change the Windows locale to be inside %APPDATA%.
It improves documentation and its easier to find (a simple ctrl+c/v from
the docs can get you there).