Remove dependency on PyScaffold during builds
As discussed in #256, after pytest-runner was deprecated and removed,
the only strong reason for having PyScaffold required during builds is
to use setuptools_scm. Docs can easily be generated via Makefile
(already created by default) or tox/nox (the new tox template already
contains a task for the docs).

A way of removing this dependency is by injecting setuptools_scm
directly on the `setup.py` template. In the future something similar can
be done for `pyproject.toml` (when it is generated by default).

This change is an attempt of doing that. As a direct consequence, we
also remove PyScaffold as a build-time dependency of itself, which makes
things easier.