Add high level sketch for changes in the API
This is an redesign attempt to solve the limitations explained in #271.
Currently the API is suffering from primitive obsession (a part of it was a
deliberate design choice to make the extension mechanism easily
understandable even for beginners).

The changes introduced here improve a bit, making the design more flexible
by replacing `FileOp` integer flags with functions.
Most of the primitive obsession is kept due to the same design decisions that
motivated the first implementation.

The following commits should ensure no errors when running PyScaffold and
add documentation.

Closes #271.