Make run_common_tasks more robust
It is important to notice that `shlex.split` will split file names when
using windows `os.sep` (`\\`). So it is required to pass a `posix=False`
as argument, otherwise we might find weird message errors, such as
"FileNotFoundError: [WinError 2] The system cannot find the file
specified" when trying to run `subprocess.check_output`.