Remove logic duplication for templates
There is some repeated logic when defining an structure with templates.
The original API accepts either strings or None objects as file
contents, but most of the time we pass either functions or
str.Templates, which leads to code duplication for the majority of the
templates.

This change intends to remove all the duplicated logic by allowing
str.Templates and callable to be passed directly as file contents,
alongside str and None (backwards compatible).