Remove mutually exclusive argparse group for extensions
This feature was never really used anyway (it existed for a particular
pair of extensions that now is outside the core - django and
cookiecutter), and particularly tricky to generalise for other use
cases... I am not even sure if it wouldn't be possible to use them at
the same time since the mutually exclusive group was implemented a while
ago (a lot have changed).

If an extension is really incompatible with other, maybe it is better to
add an action that verifies that and throw an error than during CLI args
parser (so it would also be available for raw Python API calls).