Add installation of setuptools and setuptools_scm to docs
Since editable installs still simply ignore pyproject.toml setup
dependencies and the `setup_requires` field does not play well with
pyproject.toml, build-time dependencies need to be manually installed
before running `pip install -e .`.

This commit attempts to add common instructions to the docs.
In the future, after the standardization of editable installs with
pyproject, this might not be necessary.