Attempt to implement merge_user_input using parser
Apparently not all validations from argparse happen when we call
directly `Action.__call__`, as illustrated by some of the new tests that
are failing.

Something that is also interesting to notice is that flag actions (`store_true`
or `store_false`) seem to not be called at all when not present in argv,
so calling `Action.__call__` with the other value does not yield the
expected outcome.