Merge main and run in a single function inside skeleton
As I went through #397 I found it difficult to explain/understand why we
have 2 functions in skeleton doing the CLI wrapping...

The docstring in `main` was also a bit confusing since it said that `main`
was the entry point for external calls. Since external calls can also
be interpreted as API calls from other packages, it sounded to me a bit weird that
people using skeleton's Python API should prefer passing stringfied
arguments instead of calling `fib` directly...

This change is a proposal to have `args` as an optional argument, which
makes it possible to unify the two functions into a single one, making
it a bit easier to explain.

I also added another test function for the CLI to the generated tests,
since we want to incentivise people to test.