Specify virtualenv as extra dependency for build in tox.ini
`build` seems to rely by default in the stdlib's implementation of `venv`
and unfortunately for a specific combination of Windows + conda env + virtualenv (as used by tox),
that might be problematic as identified in https://github.com/pyscaffold/pyscaffoldext-markdown/issues/10.

According to the discussion in https://github.com/pypa/build/pull/253,
specifying `virtualenv` as an extra dependency of `build` seems to be a
good workaround. Moreover I would expect some kind speed improvements
(virtualenv is reported to be faster than venv).