Remove checks for old setuptools installation
As discussed in #428, some checks we were making for the setuptools
version are incompatible with the latest version of setuptools_scm.

So one alternative (implemented in this PR) is to simply stop verifying
the versions and assume they are satisfied because they are correctly
declared as `install_requires` in `setup.cfg`.

The only doubt left is: what happens if the person is building a
generated project with `--no-pyproject`?

(In `pyproject.toml` we do have a minimum requirement for setuptools,
which is more than enough, but in `setup.py` that is omitted... Should we
add one? Is that a good practice? Or are we just being optimistic and
relying on the error message asking the user to update?)