Experiment: Change in the Sphinx Theme
Recently the RTD themes have been showing some incompatibilities with
the latest versions of Sphinx and its dependencies (see #451).

Moreover, instead of centralising the main content in
the window, the RTD theme compresses the entire page to the left, and
fills the remaining space on the right with a gray background. In large
monitors this empty are can have a side comparable (if not bigger) to
the area in which the content is displayed, which I consider unpleasant
to the reader.

This is an experiment with a new trending Sphinx theme: Furo.
It is a minimal theme which prioritises readability and have been used
lately in the documentation for the majority of the PyPA projects.

I did a series of customisations/improvements that I found suitable for
the PyScaffold docs.