Update typecheck configs for CI
This PR attempts to bring changes based on #493 for tox.ini and
.cirrus.yml (this is a long standing tradition of using mostly of the
same configuration provided by the templates to PyScaffold's own
repository and CI).

I have also noticed that, most of the times in Debian docker containers
`apt-get install` will fail unless the program is already installed,
because a `apt-get update` is required. In the Debian-based Python
containers used in Cirrus, git seem to be pre-installed.