Remove bdist_wheel section from setup.cfg template
[packaging guides] and [wheels docs] seems to
suggest that the `bdist_wheel` section is required to support
Python 2 and Python 3 at the same time. Since Python 2 was retired, it
makes sense to remove it.

PyScaffold's own `setup.cfg` file does not have this section and things
work fine (setuptools seem to default to a pure Python 3 package).

Closes #497.

[packaging guides]: https://packaging.python.org/guides/distributing-packages-using-setuptools/#wheels
[wheels docs]: https://wheel.readthedocs.io/en/stable/quickstart.html