Warn user about empty namespaces
Sometimes extension-related parameters seem to end up not persisted to
`setup.cfg`.

Issue #506, shows that it might happen for the `namespace` extension,
and via a reaction chain cause errors during updates.

The changes implemented in this commit try to prevent this reaction
chain of errors happening, while also instructing the user on how to
workaround in this situation.