Raise exception in the case extension is not installed
There are situations that the user tries to update a previously
generated project with PyScaffold, but the extensions are not installed
in the current venv (like #506).

These changes try to make this situation obvious and ask for the user to
install the missing extensions.

I believe that it is better to panic and raise an exception instead of
proceed running PyScaffold, because there are cases that the extensions
completely change the desired behaviour (e.g. namespace indirectly via
custom_extension).