Modernise test_install
There have been a bunch of changes in packaging practices for Python
since the last time we looked into `test_install`, between them:

- direct invocation of `python setup.py` is now [deprecated]
  (the existence of the file itself does not seem to be deprecated, but
  running it directly is)
- `setuptools_scm` also deprecated `python setup.py --version`
- `bdist`s, that were already deprecated, should be no longer used by
  today standards...
- PyPA is actively promoting `python -m build` as main interface

The purpose of this PR is to modernise `test_install` (and hopefully make it
faster). We do that by:

- using `python -m build {--wheel,--sdist}` instead of `python setup.py {sdist,bdist_wheel}`
- removing `bdist` tests!
- removing calls to `python setup.py --version`

I also took the opportunity to use the `venv` functions we have in place
for the `--venv` extension (reasons: more coverage for that
module + I never liked pytest-virtualenv ...).

[deprecated]: https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html