Improve off-by-one error handling in test_install
It seems that `test_install` can be flaky if running around midnight or
there is some problem with the time configuration in Python/Git.

The previous workaround worked for most situations, but not always.
For example, it could fail when the test was running at midnight of the
day 09 => 10, because these numbers differ in the first digit.

This implementation (hopefully) improves this.