Revamp `--gitlab` extension
The previous version of the template for the GitLab CI was added to
PyScaffold a while ago and never revised.

Although it seems that it was working (surprisingly), it also didn't
offer the same features as CirrusCI or the proposed GitHub Actions.

This change is a revamp on that template that tries to offer more or
less the same level of functionality our other CI configs offer.

It was tested against a replica of the `ci-tester` repository on GitLab:

- https://gitlab.com/pyscaffold/ci-tester

This is the pipeline output:

- https://gitlab.com/pyscaffold/ci-tester/-/commit/e8e3a777b972bc839a9b32d7c0c0678184ef2756