Update and simplify Cirrus CI config
The config for Cirrus CI have always been big and complicated.
While being big is not a problem (as long as it is worthy in terms of
features), it would be nice to avoid complicated configs.

I think the main culprit for the complicated config is Windows, so
lately I have been investigating ways of simplifying it.

Moreover with the latest improvements in PyScaffold in terms of
`--gitlab` and `--github-actions`, the configuration for `--cirrus` fell
a bit behind. For example, the current configuration does not support:

- aggregating coverage metrics for parallel runs
- testing the same distribution that is produced with `tox -e build`
- publishing packages to PyPI

This change bring these features to `.cirrus.yml`. Moreover the
following were also implemented:

- Added support for `freebsd` tests
- Changed default shell used for Windows to `bash` (this allows a lot of
  simplification).